==> 06-circuitscape.Rmd <==

# Circuitscape {#circuitscape}
In this chapter we use 'Circuitscape', a package built on the concepts of circuit theory, in order to explore patterns in connectivity between lowland protected areas and high elevation habitats. In contrast to the previous 'least cost paths' chapter, Circuitscape allows us to build probabilistic maps of habitat connectivity which account for the wider landscape context, rather than simply "the best route" to a given destination.

There are two import considerations when applying the Circuitscape approach:

- Given that the the surrounding landscape context is taken into account in the calculations, the issue of scale is very important: a coarse resolution map can give a very different result to a high resolution map

- High resolution maps take a huge amount of computational effort, and thus it is not possible to create a high resolution map of the whole study area.

To account for these two issues we perform the analysis at two different spatial scales. Firstly, at a coarse scale encompassing the whole study area. Then secondly, we zoom into specific locations identified as important in the least cost paths chapter to give high resolution information of the state of ridge-to-reef connectivity, and the potential for climate corridors.  


## Methods
The Circuitscape models are implemented in Julia, a platform which can process spatial data much more efficiently than R or Python. The input files are rasters, which in simple terms are three matrices of values which reflect the source points (and the electrical charge you are applying to them), destination points (and the strength of their draw) and the resistance surface (the degree to which each cell inhibits or promotes movement of charge).

In our scenario the "sources" are the low elevation protected areas, the destination points are the high elevation protected areas, and the resistance surfaces are the quality of habitats between them. 

We can then measure connectivity in two ways:

1) from the pattern of electrical current flow (i.e., animal movement probabilities) across the resistance-to-movement surface. This is the way we explore the data in this report. 

2) from the effective resistance of the circuit, measuring the degree to which the source node is isolated from the ground node. This is what is analysed in the second section. The 'effective resistance' represents a novel index which we have developed - the "Climate Resilience Index" - essentially a quantitative measure of how isolated protected areas are from future climate refugia.    


```{r, message=F, warning=F, echo=F, eval=T, include=F}
# Installed Julia according to these instructions


# https://docs.circuitscape.org/Circuitscape.jl/latest/
# https://www.mdpi.com/2073-445X/10/3/301/htm


# The descriptions in omniscape seem much better

# https://docs.circuitscape.org/Omniscape.jl/dev/usage/

library(stars)
library(sf)
library(raster)
library(leaflet)
library(terra)

merged_shp <- st_read("data/input/area_of_interest/base_aoi.shp")
#plot(st_geometry(merged_shp))
roi_17N <- st_transform(merged_shp, crs=31971)
high_pa <- st_read("data/input/area_of_interest/high_elevation_pas.shp")
high_pa_17N <- st_transform(high_pa, 31971)

#plot(st_geometry(high_pa), col="green", add=T)

simple_cost <- read_stars("Data/output/Chapter5/Low_res_resistance_surface_final.tif")
simple_cost_terra <- rast("Data/output/Chapter5/Low_res_resistance_surface_final.tif")
# Make a new folder for the synthesis chapter 7
dir.create("data/output/Chapter7")

lc_key <- read.csv("data/input/landcover/ESACCI-LC_S2_Prototype_ColorLegend.csv", header=T, sep=";")
lc_key$rgb <- rgb(lc_key$R/256,lc_key$G/256,lc_key$B/256)
# Remove no data
lc_key <- lc_key[1:11,]
# Simplify labels
lc_key$LCCOwnLabel[lc_key$LCCOwnLabel=="Tree cover areas"] <- "Tree cover"
lc_key$LCCOwnLabel[lc_key$LCCOwnLabel=="Shrubs cover areas"] <- "Shrub cover"
lc_key$LCCOwnLabel[lc_key$LCCOwnLabel=="Vegetation aquatic or regularly flooded"] <- "Flooded vegetation"



```

## Electrical current flow (a.k.a. animal movement)

### Coarse scale (1km)

#### Current proportional to protected area size
For our first analysis, we apply a charge relative to the size of each of the lowland protected areas - the larger the protected area the higher the electrical current applied. Current is free to move to any of the high elevation protected areas within the landscape. 

Isolated, small lowland protected areas will only contribute small amounts of charge. Large interconnected protected areas with contribute large amounts of charge. 

Below is the output from this analysis across the whole survey area. The colors denote the magnitude of current flow: blue = low current flow, yellow = high current flow. The black polygons represent the destination high elevation protected areas. 

```{r, eval=F, echo=F, message=F, include=F}

# Import start polygons
lowland_pa <- st_read("data/input/protected_areas/lowland_protected_areas.shp")

leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  #addRasterImage(tmp, group="Raster") %>% 
  addPolygons(data=lowland_pa, fillColor = "yellow",
               label = lowland_pa$NAME, group="lowland_pa") %>% 
  #addCircleMarkers(data=end_points_wgs, col="black", radius=1) %>% 
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Costal_pa"))

lowland_pa$land_size <- as.numeric(st_area(lowland_pa)/(1000*1000))

# Transform to 17N
start_poly_17N <- st_transform(lowland_pa, 31971)
st_write(start_poly_17N, "Data/output/Chapter5/start_polygons_17N.shp", append=FALSE)
st_write(lowland_pa, "Data/output/Chapter5/start_polygons.shp", append=FALSE)

# starting parks
leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addPolygons(data=lowland_pa, fillColor = "yellow",
               label = lowland_pa$NAME, group="starting_points") %>%
  #addPolygons(data=lowland_pa, fillColor = "red",
  #             label = lowland_pa$NAME, group="unfiltered") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("starting_points", "unfiltered"))
 


 


# Create folders to store the results
dir.create("Data/output/Chapter6/lowland_pa_all_low_res/")
path_text <- "Data/output/Chapter6/lowland_pa_all_low_res/"

# Import the costa layer
simple_cost <- read_stars("Data/output/Chapter5/Low_res_resistance_surface_final.tif")
plot(simple_cost)

simple_ras <- raster("Data/output/Chapter5/Low_res_resistance_surface_final.tif")

# Plot the resistance surface (check rivers not impassible)
leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(simple_ras, group="Raster") #%>% 
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Raster")) #%>%
 

# Start points
start_point <- st_point_on_surface(lowland_pa)
start_point_17N <- st_transform(start_point, 31971)

# Read in the final destinations
end_poly <- st_read("data/input/area_of_interest/high_elevation_pas.shp")
end_poly_17N <- st_transform(end_poly, 31971)
# Split into individual polygons
end_poly_17N_filtered <- st_cast(end_poly_17N,"POLYGON")
# Add new area
end_poly_17N_filtered$fragArea <- as.numeric(st_area(end_poly_17N_filtered)/(1000*1000))

# Remove small fragments
end_poly_17N_filtered <- end_poly_17N_filtered[end_poly_17N_filtered$fragArea>=1,]
nrow(end_poly_17N_filtered)
# End points
end_point <- st_point_on_surface(end_poly_17N_filtered)
end_point_17N <- st_point_on_surface(end_poly_17N_filtered)
nrow(end_point_17N)

##############

tmp <- project(simple_cost_terra, crs(start_point_17N))
writeRaster(tmp,"Data/output/Chapter5/Low_res_resistance_surface_final_17N.tif" , overwrite=T)

cellmap <- raster("Data/output/Chapter5/Low_res_resistance_surface_final.tif")
NAvalue(cellmap) <- -9999
plot(cellmap)
writeRaster(cellmap, paste0(path_text,"/cellmap.asc"), overwrite=T)

#############
## Starting layers (terrestrial protected areas)
#start_blocks <- st_intersection(terrestrial_node_17N, aoi_test) 
#start_blocks <- start_poly_17N

lc <- rast("Data/output/Chapter5/Low_res_resistance_surface_final.tif")
start <- vect("Data/output/Chapter5/start_polygons.shp")
start_sf <- st_read("Data/output/Chapter5/start_polygons_17N.shp")
end <- vect("data/input/area_of_interest/high_elevation_pas.shp")

start_sp <- as(start, "Spatial")
start_v <- vect(start_sp)
end_tmp <- as(end, "Spatial")
end_v <- vect(end_tmp)
start$area <- round(start_v$area,2)

end_ras <- rasterize(end,lc, field=999)
polygons <- end_ras
#plot(polygons)

# This takes a centroid of each protected area
start_p <- centroids(start_v, inside=T)

spaced_points_17N <- st_read("Data/output/Chapter5/start_points_1km_spacing.shp")
# Convert to points
spaced_points_17N <- st_cast(spaced_points_17N, "POINT")
# Convert to WGS
spaced_points_wgs <- st_transform(spaced_points_17N, 4327)
#Plot in leaflet
# starting parks
leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addPolygons(data=start_sp, fillColor = "yellow",
               label = start_sp$NAME, group="starting_points") %>%
  addCircleMarkers(data=spaced_points_wgs, col="black", radius=1) %>% 
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("starting_points", "unfiltered")) %>%
  addScaleBar()
  
sources <- rasterize(spaced_points_wgs,lc, 
             field="area") # make the 

end_p <- centroids(end_v, inside=T)

grounds <- rasterize(end_p,lc, 
             field=1) # make the area the current

plot(cost_final_ras)
plot(start_v, add=T)
plot(spaced_points_wgs, add=T)

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

plot(polygons_ras)
plot(sources_ras)
plot(grounds_ras)

# I used this version when I used opne start point per protected areas, and weighted the input curtrent based on protected area size - see below for normalised version
#writeRaster(sources_ras, paste0(path_text,"sources.asc"), overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, paste0(path_text, "grounds.asc"), overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, paste0(path_text,"regions_grid.asc"), overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# And a normalised version
sources_normal <- sources_ras
sources_normal[sources_normal>0]<-1
plot(sources_normal)

writeRaster(sources_normal, paste0(path_text,"/sources.asc"), overwrite=T)

dir.create(paste0(path_text,"/output"))


######################
# Run in JULIA




# JULIA CODE TO RUN (install circuitscape first)

# > using Circuitscape
# > cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\lowland_pa_all_low_res")
# > compute("pas_lowres.ini")

# Runs 30 seconds on my machine - 8 parallel processes

```

```{r, echo=F, message=F, warning=F, include=F}
# Plot the result
res_cur <- rast("data/output/Chapter6/lowland_pa_all_low_res/output/pas_lowres_curmap.asc")
# GIve it the crs
terra::crs(res_cur) <- "EPSG:4326"

res_volt <- rast("data/output/Chapter6/lowland_pa_all_low_res/output/pas_lowres_voltmap.asc")

end_poly <- vect("data/input/area_of_interest/high_elevation_pas.shp")
#end_poly_17N <- project(end_poly, "EPSG:31971")
#plot(end_poly)

# Raw current
#plot(res_cur)
res_clamp <- clamp(res_cur, upper=100)
# Make low current NA
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.0001] <- NA; return(x)} )

# Masked current (no sources)
res_clamp <- mask(res_clamp, end_poly, inverse=T)

# Check before convery to raster
#plot(res_clamp)
#plot(st_geometry(roi_17N), add=T)

pa_cond_final <- as(res_clamp, "Raster")

# Reimport start and ends
end_sf <- st_read("data/input/area_of_interest/high_elevation_pas.shp")
#end_sf_17N <- st_transform(end_sf, 31971)
# pdf("data/images/pa_cur_low.pdf",height=6.5, width=8)
# plot(pa_cond_final)
# plot(st_geometry(roi_17N), add=T)
# plot(st_geometry(end_poly_17N), add=T, col="black")
# dev.off()


writeRaster(pa_cond_final, "data/output/Chapter7/Conductivity_pa_high_100m_res.tif", overwrite=T)
```


```{r, echo=F, message=F, warning=F}
# Define a palette
pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(pa_cond_final),1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(pa_cond_final, group="Raster", colors = pal) %>% 
  addPolygons(data=end_sf, fillColor = "red",
               label = end_sf$NAME, group="High_pa", stroke=F, fillOpacity = 1) %>% 
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Raster","High_pa")) %>%
  addLegend(pal = pal, values = values(pa_cond_final),
    title = "Current flow")      #%>%
  #addPolylines(data=tmp_lcp, weight = 2, fillColor = "blue",
   #            label = tmp_lcp$ID)# %>% 
  #addCircleMarkers(data=tmp_end, col="black", radius=1)

writeRaster(pa_cond_final, "data/output/Chapter7/central_america_connectivity_pa_size.tiff", overwrite=T)
writeRaster(pa_cond_final, "data/output/Chapter6/Climate_hub_current_and_resistance/Central_america_current.tif", overwrite=T)


lc <- rast("Data/output/Chapter5/Low_res_resistance_surface_final.tif")


writeRaster(lc, "data/output/Chapter6/Climate_hub_current_and_resistance/Central_america_conductance.tif", overwrite=T)

```


```{r, eval=F}
# Map for board presentation
s <- st_read("data/output/Chapter7/draft_climate_hubs.shp")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(pa_cond_final, group="Raster", colors = pal) %>% 
  addPolygons(data=s, fillColor = "red",stroke=T, color="yellow",
               label = s$id, group="High_pa")
```

The map reveals several key conclusions:

i) The Caribbean coast his a much higher degree of current flow (connectivity) than the pacific coast. This is unsurprising as the protected areas on the Caribbean coast are typically larger, and less disturbed than on the Pacific coast (refer to the human modification maps and protected areas maps in Chapter 2.

ii) Not only is the absolute magnitude of current flow higher on the Caribbean coast, it is also less constrained - there are more available routes to animals and on the Pacific Coast. Contrast those patterns to the north west of the map in El Salvador and Guatemala - all of the charge is being funneled through one narrow corridor. This represents a bottleneck in animal movement. 

iii) The patterns in current flow closely mirror the availability of high quality forest habitat. 

It is important to not that the exact flow of current should be treated with caution as the map resolution is very coarse. See below for a high resolution case study. 

#### Equal current
Although the above method is good to understand how the number of animals flows across the landscape, if we are directly interested in the paths themselves we can give each protected area the same amount of electrical current.  We then more easily compare patterns at the landscape scale.

```{r, eval=T, echo=F, message=F, warning=F}
# Normalised source
res_cur_norm <- rast("data/output/Chapter6/Coastal_pa_all_low_res/output/pas_lowres_normalised_curmap.asc")
#terra::crs(res_cur_norm) <- "EPSG:31971"

res_clamp <- clamp(res_cur_norm, upper=0.5)
# Make low current NA
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.00001] <- NA; return(x)} )
#plot(res_clamp)

res_cur_norm_mask <- mask(res_clamp, end_poly, inverse=T)


#And the fragment importance like this:
res_cur_norm_ras <- raster(res_clamp)

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), values(res_cur_norm_mask),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(res_cur_norm_ras, group="Raster", colors = pal) %>%  #  
  addPolygons(data=high_pa, fillColor = "black",
               label = high_pa$NAME, group="High_pa", stroke=F, fillOpacity = 1) %>% 
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Raster","High_pa")) %>%
  addLegend(pal = pal, values = values(res_cur_norm_mask),
    title = "Current flow")      #%>%
  #addPolylines(data=tmp_lcp, weight = 2, fillColor = "blue",
   #            label = tmp_lcp$ID)# %>% 
  #addCircleMarkers(data=tmp_end, col="black", radius=1)


writeRaster(res_cur_norm_mask, "data/output/Chapter7/central_america_connectivity_normalised.tiff", overwrite=T)
```


#### High elevation area importance
A third way in which we can use Circuitscape to understand patterns in animal connectivity is the hold all of the source locations at a constant voltage, then quantify the amount of current each of the destination locations accrues. This will give an index of the "importance" of each high elevation protected area: the higher the score it accumulates the higher the number of lowland protected areas which are contributing charge towards it, and the more "important" the high elevation protected area to landscape scale climate resilience. 


```{r, echo=F, message=F, warning=F, include=F}
# Normalised source
res_cur_norm <- raster(res_cur_norm)
res_cur_norm_mask_2 <- mask(res_cur_norm, high_pa, inverse=F)
plot(res_cur_norm_mask_2)

# Define a palette
pal <- colorNumeric(c("#91bfdb", "#ffffbf", "#fc8d59"), values(res_cur_norm_mask_2),
  na.color = "transparent")
crs(res_cur_norm_mask_2) <- "EPSG:4326"

#plot(res_cur_norm_mask_2)
```


```{r, echo=F, message=F, warning=F}

leaflet() %>%
  #addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldGrayCanvas") %>%
  addRasterImage(res_cur_norm_mask_2, group="Raster", colors = pal) %>% 
  addPolygons(data=end_sf, fillColor = "black",
               label = end_sf$NAME, group="High_pa", stroke=F, fillOpacity = 0) %>% 
  addLegend(pal = pal, values = values(res_cur_norm_mask_2),
    title = "Current accrued")      #%>%
  #addPolylines(data=tmp_lcp, weight = 2, fillColor = "blue",
   #            label = tmp_lcp$ID)# %>% 
  #addCircleMarkers(data=tmp_end, col="black", radius=1)


```


This plot demonstrates that La Amistad receives input from a disproportionately high number of lowland protected areas. Other important protected areas are Sierra de las Minas in Guatamala, and Cordillera Volcanica Central in Costa Rica.  


### Fine scale (100m)

The maps shown above are useful, but they are at a scale which is not amenable to making explicit management recommendations. The next step is to use a higher resolution cost surface, see [Chapter 5](#cost-surface_high).

From earlier model tests we determined that we can run landscapes at 100m x 100m resolutions in two hours. 10 x 10m grids would take days. I now provide an example case study based on the Osa Peninsula region of the power of high resolution connectivity analyses.

```{r, include=F, eval=T, echo=F}
# Load data common to all locations

library(terra)
cost_final<- rast("Data/output/Chapter5/High_res_resistance_surface_final.tif")

#Subset to just osa landscape xmin, x max, ymin, ymax
hubs <- vect("data/output/Chapter7/draft_climate_hubs.shp")

#Start and end polygons
start_all <- vect("data/input/protected_areas/lowland_protected_areas.shp")
end_all <- vect("data/input/area_of_interest/high_elevation_pas.shp")

# Start points
start_v <- vect("Data/output/Chapter5/start_points_1km_spacing.shp") #17N
start_v_wgs <- project(start_v, lc)


# MAke a folder to store the final (clamped) values for sharing
dir.create("Data/output/Chapter6/Climate_hub_current_and_resistance")

```




#### Case Study 1: Osa Peninsula

Here we use the high resolution resistance layer developed in Chapter 4 to quantify finescale patterns in connectivity within the Osa Peninsula. As in the central america wide example, each protected area contributes an electrical current proportion to the size of the park. Current is applied at the midpoint (rather than the protected area as a whole) to allow current to move naturally throughout the protected areas.

As above, the colors denote the magnitude of current flow: blue = low current flow, yellow = high current flow. 

```{r, include=F, eval=F, echo=F}

##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
osa <- hubs[hubs$ID=="Amistosa",]
aoi <-  ext(osa)
cost_sub  <- crop(cost_final, aoi)

cost_final_ras <- as(cost_sub,"Raster")

leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_1_Osa/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_1_Osa/cost_surface_100m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_1_Osa/cost_surface_100m.tif")
# lets try a high res circuit scape  
#start <- focal_pa[focal_pa$NAME=="Corcovado",]
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)


# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
start_p<- tmp[crds(tmp)[,1]<(-83),]

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)
grounds <- rasterize(end_p,lc, 
             field=1) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")


# Write the objects
dir.create("Data/output/Chapter6/Case_study_1_Osa")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_1_Osa/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_1_Osa/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_1_Osa/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_1_Osa/output")

# Julia codes

#cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_2")
#compute("case_study_2.ini")

# Time taken... 10 seconds we could do 10m res?



```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_1_Osa/cost_surface_100m.tif")
aoi <- ext(lc)

res_cur <- rast("Data/output/Chapter6/Case_study_1_Osa/output/Case_study_1_curmap.asc")

end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur, zlim=c(0.1,15))

res_clamp <- clamp(res_cur, upper=15)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.1] <- NA; return(x)} )

# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
cost    <- as(lc,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),15.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost, group="Conductance") %>%
  addRasterImage(final_cur, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 

# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/Osa_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/Osa_conductance.tif", overwrite=TRUE, NAflag=-9999)

```

The above map shows two interesting elements:

i) Firstly, it highlights the importance of protected and contiguous habitat in the neck of the peninsula, north of Corcovado. All routes must pass through here, hence the electrical current flow is relatively high.

ii) Secondly, despite the existence of existing "corridors" through the matrix of disturbed and human modified habitat between the peninsula and La Amistad to the north - there is no clear patch of least resistance. Cnnectivity appears to be highly diffuse and low across the whole region. This suggests that there is only weak functional connectivity between the Peninsula and La Amistad, and that climate resilience may be low. 

It is also import to check if connectivity map generated holds up local scales. An example plantation below suggests that the electric flow maps well on to how we would expect large terrestrial mammals to move through this system:

```{r , echo=F, message=F, warning=F}
knitr::include_graphics("data/images/example.PNG")
```

Which suggests that such maps can be used to direct local restoration efforts aimed at improving ridge to reef connectivity in the region.

#### Case Study 2: Guanacaste
This case study details the connections between Santa Rosa National Park and the highland areas of Volcan Cacao, Rincon de la Vieja and National Park Miravalles - Jorge Manuel Dengo. All of the highland areas have some degree of connectivity, however there is a very important corridor of connectivity (hot yellow in color) between Santa Rosa and the highlands of Volcan Cacao (see below). Much of this landscape is protected (to some degree), so this could be an example of a relatively "safe" climate corridor. 

```{r, include=F, eval=F, echo=F}
##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
guanacaste <- hubs[hubs$ID=="Guancaste",]
aoi <-  ext(guanacaste)
cost_sub  <- crop(cost_final, aoi)

cost_final_ras <- as(cost_sub,"Raster")

leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_2/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_2/cost_surface_100m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_2/cost_surface_100m.tif")
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)


# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
plot(tmp)
start_p<- tmp[crds(tmp)[,1]<(-85.45) & crds(tmp)[,2]>10.6,]
guanacaste <- hubs[hubs$ID=="Guancaste",]
#plot(guanacaste)
#plot(tmp, add=T, col="red")
#plot(start_p,add=T)

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)
grounds <- rasterize(end_p,lc, 
             field=1) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

plot(sources_ras)
# Write the objects
dir.create("Data/output/Chapter6/Case_study_2")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_2/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_2/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_2/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_2/output")

# Julia codes

# cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_2")
# compute("case_study_2.ini")

# Time taken... 10 seconds we could do 10m res?



```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_2/cost_surface_100m.tif")
aoi <- ext(lc)

# Plot the result - TIME TAKEN 6454 seconds - will try with different solver
res_cur <- rast("Data/output/Chapter6/Case_study_2/output/Case_study_1_curmap.asc")

end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur)

res_clamp <- clamp(res_cur, upper=10)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.1] <- NA; return(x)} )
#plot(res_clamp)
# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
cost    <- as(lc,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),15.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost, group="Conductance") %>%
  addRasterImage(final_cur, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 



# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/Guanacaste_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/Guanacaste_conductance.tif", overwrite=TRUE, NAflag=-9999)

```


#### Case Study 3: Selva Maya

```{r, include=F, eval=F, echo=F}
##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
selva <- hubs[hubs$ID=="SelvaMaya",]
aoi <-  ext(selva)
cost_sub  <- crop(cost_final, aoi)

# Down sample to plot - otherwise it is too large
cost_sub_low <- aggregate(cost_sub, fact=4, fun="mean", na.rm=T)

cost_final_ras <- as(cost_sub,"Raster")
cost_final_ras_low <- as(cost_sub_low,"Raster")
plot(cost_sub)
plot(selva, add=T)

leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras_low, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_3/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_3/cost_surface_100m.tif", overwrite=TRUE)
writeRaster(cost_final_ras_low, "Data/output/Chapter6/Case_study_3/cost_surface_300m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_3/cost_surface_100m.tif")
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)


# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
plot(tmp)
start_p<- tmp 
selva <- hubs[hubs$ID=="SelvaMaya",]
#plot(selva)
#plot(tmp, add=T, col="red")
#plot(start_p,add=T)

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)
grounds <- rasterize(end_p,lc, 
             field=1) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

#plot(sources_ras)
#plot(polygons_ras,add=T)

# Write the objects
dir.create("Data/output/Chapter6/Case_study_3")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_3/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_3/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_3/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_3/output")

# Julia codes

# cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_3")
# compute("case_study_3.ini")

# Time taken... 10 seconds we could do 10m res?



```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_3/cost_surface_100m.tif")
lc_low <- rast("Data/output/Chapter6/Case_study_3/cost_surface_300m.tif")

aoi <- ext(lc)

# Plot the result - TIME TAKEN 6454 seconds - will try with different solver
res_cur <- rast("Data/output/Chapter6/Case_study_3/output/Case_study_3_curmap.asc")

end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur)

res_clamp <- clamp(res_cur, upper=40)
#plot(res_clamp)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.1] <- NA; return(x)} )
#plot(res_clamp)
# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

# I am going to have to downscalew this too
res_clamp_low <- aggregate(res_clamp, fact=4, fun="mean", na.rm=T)


#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
final_cur_low    <- as(res_clamp_low,"Raster")

cost    <- as(lc,"Raster")
cost_low    <- as(lc_low,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),40.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost_low, group="Conductance") %>%
  addRasterImage(final_cur_low, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 


# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/SelvaMaya_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/SelvaMaya_conductance.tif", overwrite=TRUE, NAflag=-9999)

```

#### Case Study 4: Belize

```{r, include=F, eval=F, echo=F}
##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
belize <- hubs[hubs$ID=="Belize",]
aoi <-  ext(belize)
cost_sub  <- crop(cost_final, aoi)

# Down sample to plot - otherwise it is too large
cost_sub_low <- aggregate(cost_sub, fact=4, fun="mean", na.rm=T)

cost_final_ras <- as(cost_sub,"Raster")
cost_final_ras_low <- as(cost_sub_low,"Raster")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras_low, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_4/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_4/cost_surface_100m.tif", overwrite=TRUE)
writeRaster(cost_final_ras_low, "Data/output/Chapter6/Case_study_4/cost_surface_300m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_4/cost_surface_100m.tif")
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)

# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
plot(tmp)
start_p<- tmp#[crds(tmp)[,1]<(-85.45) & crds(tmp)[,2]>10.6,]
belize <- hubs[hubs$ID=="Belize",]
#plot(belize)
#plot(tmp, add=T, col="red")
#plot(start_p,add=T)

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)
grounds <- rasterize(end_p,lc, 
             field=1) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

plot(sources_ras)
# Write the objects
dir.create("Data/output/Chapter6/Case_study_4")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_4/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_4/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_4/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_4/output")

# Julia codes

# cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_4")
# compute("case_study_4.ini")

# Time taken... 10 seconds we could do 10m res?



```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_4/cost_surface_300m.tif")
aoi <- ext(lc)

# Plot the result - TIME TAKEN 6454 seconds - will try with different solver
res_cur <- rast("Data/output/Chapter6/Case_study_4/output/Case_study_4_curmap.asc")

end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur)

res_clamp <- clamp(res_cur, upper=40)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.1] <- NA; return(x)} )
#plot(res_clamp)

# I am going to have to downscalew this too
res_clamp_low <- aggregate(res_clamp, fact=4, fun="mean", na.rm=T)

# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
final_cur_low    <- as(res_clamp_low,"Raster")

cost    <- as(lc,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),40.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost, group="Conductance") %>%
  addRasterImage(final_cur_low, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 


# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/Belize_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/Belize_conductance.tif", overwrite=TRUE, NAflag=-9999)


```


#### Case Study 5: Mostikia

```{r, include=F, eval=F, echo=F}
##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
mostikia <- hubs[hubs$ID=="Mostikia",]
aoi <-  ext(mostikia)
cost_sub  <- crop(cost_final, aoi)

# Down sample to plot - otherwise it is too large
cost_sub_low <- aggregate(cost_sub, fact=4, fun="mean", na.rm=T)

cost_final_ras <- as(cost_sub,"Raster")
cost_final_ras_low <- as(cost_sub_low,"Raster")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras_low, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_5/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_5/cost_surface_100m.tif", overwrite=TRUE)
writeRaster(cost_final_ras_low, "Data/output/Chapter6/Case_study_5/cost_surface_300m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_5/cost_surface_100m.tif")
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)

# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
plot(tmp)
start_p<- tmp#[crds(tmp)[,1]<(-85.45) & crds(tmp)[,2]>10.6,]
mostika <- hubs[hubs$ID=="Mostikia",]
#plot(mostika)
#plot(tmp, add=T, col="red")
#plot(start_p,add=T)

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)

tmp <- st_as_sf(end_p)
#plot(tmp,add=T)

grounds <- rasterize(tmp,lc, 
             field=20) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

plot(sources_ras)
plot(polygons_ras)
plot(grounds_ras)

# Write the objects
dir.create("Data/output/Chapter6/Case_study_5")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_5/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_5/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_5/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_5/output")

# Julia codes

# cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_5")
# compute("case_study_5.ini")

```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_5/cost_surface_300m.tif")
aoi <- ext(lc)

# Plot the result - TIME TAKEN 6454 seconds - will try with different solver
res_cur <- rast("Data/output/Chapter6/Case_study_5/output/Case_study_5_curmap.asc")

end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur)

res_clamp <- clamp(res_cur, upper=40)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.1] <- NA; return(x)} )

# I am going to have to downscale this too
res_clamp_low <- aggregate(res_clamp, fact=3, fun="mean", na.rm=T)


#plot(res_clamp)
# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
final_cur_low    <- as(res_clamp_low,"Raster")

cost    <- as(lc,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),40.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost, group="Conductance") %>%
  addRasterImage(final_cur, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 


# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/Mostikia_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/Mostikia_conductance.tif", overwrite=TRUE, NAflag=-9999)


```




#### Case Study 6: IndioMaiz

```{r, include=F, eval=F, echo=F}
##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
indio <- hubs[hubs$ID=="IndioMaiz",]
aoi <-  ext(indio)
cost_sub  <- crop(cost_final, aoi)

# Down sample to plot - otherwise it is too large
cost_sub_low <- aggregate(cost_sub, fact=2, fun="mean", na.rm=T)

cost_final_ras <- as(cost_sub,"Raster")
cost_final_ras_low <- as(cost_sub_low,"Raster")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras_low, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_6/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_6/cost_surface_100m.tif", overwrite=TRUE)
writeRaster(cost_final_ras_low, "Data/output/Chapter6/Case_study_6/cost_surface_200m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_6/cost_surface_100m.tif")
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)

# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
plot(tmp)
start_p<- tmp#[crds(tmp)[,1]<(-85.45) & crds(tmp)[,2]>10.6,]
indio <- hubs[hubs$ID=="IndioMaiz",]
#plot(indio)
#plot(tmp, add=T, col="red")
#plot(start_p,add=T)

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)
tmp <- st_as_sf(end_p)
#plot(tmp,add=T)

grounds <- rasterize(tmp,lc, 
             field=1) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

plot(sources_ras)
plot(polygons_ras)
plot(grounds_ras)

# Write the objects
dir.create("Data/output/Chapter6/Case_study_6")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_6/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_6/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_6/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_6/output")

# Julia codes

# cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_6")
# compute("case_study_6.ini")

```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_6/cost_surface_200m.tif")
aoi <- ext(lc)

# Plot the result - TIME TAKEN 6454 seconds - will try with different solver
res_cur <- rast("Data/output/Chapter6/Case_study_6/output/Case_study_6_curmap.asc")
  
end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur)

res_clamp <- clamp(res_cur, upper=40)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.1] <- NA; return(x)} )

# I am going to have to downscale this too
res_clamp_low <- aggregate(res_clamp, fact=2, fun="mean", na.rm=T)


#plot(res_clamp)
# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
final_cur_low    <- as(res_clamp_low,"Raster")

cost    <- as(lc,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),40.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost, group="Conductance") %>%
  addRasterImage(final_cur_low, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 

# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/IndioMaiz_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/IndioMaiz.tif", overwrite=TRUE, NAflag=-9999)


```



#### Case Study 7: Amistad

```{r, include=F, eval=F, echo=F}
##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
amistad <- hubs[hubs$ID=="Amistad",]
aoi <-  ext(amistad)
cost_sub  <- crop(cost_final, aoi)

# Down sample to plot - otherwise it is too large
#cost_sub_low <- aggregate(cost_sub, fact=2, fun="mean", na.rm=T)

cost_final_ras <- as(cost_sub,"Raster")
#cost_final_ras_low <- as(cost_sub_low,"Raster")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_7/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_7/cost_surface_100m.tif", overwrite=TRUE)
#writeRaster(cost_final_ras_low, "Data/output/Chapter6/Case_study_7/cost_surface_200m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_7/cost_surface_100m.tif")
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)

# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
plot(tmp)
start_p<- tmp#[crds(tmp)[,1]<(-85.45) & crds(tmp)[,2]>10.6,]
amistad <- hubs[hubs$ID=="Amistad",]
#plot(amistad)
#plot(tmp, add=T, col="red")
#plot(start_p,add=T)

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)
tmp <- st_as_sf(end_p)
#plot(tmp,add=T)

grounds <- rasterize(tmp,lc, 
             field=1) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

plot(sources_ras)
plot(polygons_ras)
plot(grounds_ras)

# Write the objects
dir.create("Data/output/Chapter6/Case_study_7")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_7/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_7/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_7/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_7/output")

# Julia codes

# cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_7")
# compute("case_study_7.ini")

```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_7/cost_surface_100m.tif")
aoi <- ext(lc)

# Plot the result - TIME TAKEN 6454 seconds - will try with different solver
res_cur <- rast("Data/output/Chapter6/Case_study_7/output/Case_study_7_curmap.asc")
  
end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur)

res_clamp <- clamp(res_cur, upper=3)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.1] <- NA; return(x)} )

# I am going to have to downscale this too
#res_clamp_low <- aggregate(res_clamp, fact=2, fun="mean", na.rm=T)


#plot(res_clamp)
# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
#final_cur_low    <- as(res_clamp_low,"Raster")

cost    <- as(lc,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),3.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost, group="Conductance") %>%
  addRasterImage(final_cur, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 

# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/Amistad_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/Amistad_conductance.tif", overwrite=TRUE, NAflag=-9999)


```



#### Case Study 8: Darien

```{r, include=F, eval=F, echo=F}
##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
darien <- hubs[hubs$ID=="Darien",]
aoi <-  ext(darien)
cost_sub  <- crop(cost_final, aoi)

# Down sample to plot - otherwise it is too large
cost_sub_low <- aggregate(cost_sub, fact=2, fun="mean", na.rm=T)

cost_final_ras <- as(cost_sub,"Raster")
cost_final_ras_low <- as(cost_sub_low,"Raster")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras_low, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_8/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_8/cost_surface_100m.tif", overwrite=TRUE)
writeRaster(cost_final_ras_low, "Data/output/Chapter6/Case_study_8/cost_surface_200m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_8/cost_surface_100m.tif")
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)

# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
plot(tmp)
start_p<- tmp#[crds(tmp)[,1]<(-85.45) & crds(tmp)[,2]>10.6,]
amistad <- hubs[hubs$ID=="Amistad",]
#plot(amistad)
#plot(tmp, add=T, col="red")
#plot(start_p,add=T)

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)
tmp <- st_as_sf(end_p)
#plot(tmp,add=T)

grounds <- rasterize(tmp,lc, 
             field=1) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

plot(sources_ras)
plot(polygons_ras)
plot(grounds_ras)

# Write the objects
dir.create("Data/output/Chapter6/Case_study_8")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_8/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_8/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_8/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_8/output")

# Julia codes

# cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_8")
# compute("case_study_8.ini")

```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_8/cost_surface_200m.tif")
aoi <- ext(lc)

# Plot the result - TIME TAKEN 6454 seconds - will try with different solver
res_cur <- rast("Data/output/Chapter6/Case_study_8/output/Case_study_8_curmap.asc")
  
end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur)

res_clamp <- clamp(res_cur, upper=10)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.1] <- NA; return(x)} )

# I am going to have to downscale this too
res_clamp_low <- aggregate(res_clamp, fact=2, fun="mean", na.rm=T)


#plot(res_clamp)
# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
final_cur_low    <- as(res_clamp_low,"Raster")

cost    <- as(lc,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),10.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost, group="Conductance") %>%
  addRasterImage(final_cur_low, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 


# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/Darien_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/Darien_conductance.tif", overwrite=TRUE, NAflag=-9999)



```


#### Case Study 9: El Salvador

```{r, include=F, eval=F, echo=F}
##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
salva <- hubs[hubs$ID=="ElSalvador",]
aoi <-  ext(salva)
cost_sub  <- crop(cost_final, aoi)

# Down sample to plot - otherwise it is too large
cost_sub_low <- aggregate(cost_sub, fact=2, fun="mean", na.rm=T)

cost_final_ras <- as(cost_sub,"Raster")
cost_final_ras_low <- as(cost_sub_low,"Raster")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras_low, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_9/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_9/cost_surface_100m.tif", overwrite=TRUE)
writeRaster(cost_final_ras_low, "Data/output/Chapter6/Case_study_9/cost_surface_200m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_9/cost_surface_100m.tif")
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)

# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
plot(tmp)
start_p<- tmp#[crds(tmp)[,1]<(-85.45) & crds(tmp)[,2]>10.6,]
salva <- hubs[hubs$ID=="ElSalvador",]
#plot(salva)
#plot(tmp, add=T, col="red")
#plot(start_p,add=T)

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)
tmp <- st_as_sf(end_p)
#plot(tmp,add=T)

grounds <- rasterize(tmp,lc, 
             field=1) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

plot(sources_ras)
plot(polygons_ras)
plot(grounds_ras)

# Write the objects
dir.create("Data/output/Chapter6/Case_study_9")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_9/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_9/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_9/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_9/output")

# Julia codes

# cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_9")
# compute("case_study_9.ini")

```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_9/cost_surface_200m.tif")
aoi <- ext(lc)

# Plot the result - TIME TAKEN 6454 seconds - will try with different solver
res_cur <- rast("Data/output/Chapter6/Case_study_9/output/Case_study_9_curmap.asc")
  
end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur)

res_clamp <- clamp(res_cur, upper=5)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.01] <- NA; return(x)} )

# I am going to have to downscale this too
res_clamp_low <- aggregate(res_clamp, fact=2, fun="mean", na.rm=T)


#plot(res_clamp)
# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
final_cur_low    <- as(res_clamp_low,"Raster")

cost    <- as(lc,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),5.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost, group="Conductance") %>%
  addRasterImage(final_cur_low, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 


# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/ElSalvador_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/ElSalvador_conductance.tif", overwrite=TRUE, NAflag=-9999)



```



#### Case Study 10: Honduras Volcan

```{r, include=F, eval=F, echo=F}
##### PREP THE HIGH RESOLUTION LAYER FOR STEP 1
hond <- hubs[hubs$ID=="HondurasVolcan",]
aoi <-  ext(hond)
cost_sub  <- crop(cost_final, aoi)

# Down sample to plot - otherwise it is too large
cost_sub_low <- aggregate(cost_sub, fact=2, fun="mean", na.rm=T)

cost_final_ras <- as(cost_sub,"Raster")
cost_final_ras_low <- as(cost_sub_low,"Raster")


leaflet() %>%
  addProviderTiles("OpenStreetMap", group="OSM (default)") %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addRasterImage(cost_final_ras_low, group="Cost (100m)") %>%
  addLayersControl(
    baseGroups = c("OSM (default)", "Satellite"),
    overlayGroups = c("Cost (100m)")) 


writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_10/cellmap.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)
writeRaster(cost_final_ras, "Data/output/Chapter6/Case_study_10/cost_surface_100m.tif", overwrite=TRUE)
writeRaster(cost_final_ras_low, "Data/output/Chapter6/Case_study_10/cost_surface_200m.tif", overwrite=TRUE)

```


```{r, echo=F, eval=F, message=F, warning=F}
lc <- rast("Data/output/Chapter6/Case_study_10/cost_surface_100m.tif")
aoi <- ext(lc) 

#start_ras <- rasterize(start_v,lc_re_small, field=1)
end_ras <- rasterize(end_all,lc, field=999)
polygons <- end_ras
plot(polygons)

# Intersect with our starting points layer
#Import points

# Remove locations outside of focal area
tmp <- crop(start_v_wgs, aoi)
plot(tmp)
start_p<- tmp#[crds(tmp)[,1]<(-85.45) & crds(tmp)[,2]>10.6,]
hond <- hubs[hubs$ID=="HondurasVolcan",]
#plot(hond)
#plot(tmp, add=T, col="red")
#plot(start_p,add=T)

# Create source layer
sources <- rasterize(start_p,lc) # make the 

# Create end layer
end_v <- crop(end_all, aoi)
end_p <- centroids(end_v, inside=T)
tmp <- st_as_sf(end_p)
#plot(tmp,add=T)

grounds <- rasterize(tmp,lc, 
             field=1) # make the area the current

lc    <- as(lc,"Raster")
polygons_ras <- as(polygons,"Raster")
sources_ras    <- as(sources,"Raster")
grounds_ras    <- as(grounds,"Raster")

plot(sources_ras)
plot(polygons_ras)
plot(grounds_ras)

# Write the objects
dir.create("Data/output/Chapter6/Case_study_10")

# Source
writeRaster(sources_ras, "Data/output/Chapter6/Case_study_10/sources.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Ground
writeRaster(grounds_ras, "Data/output/Chapter6/Case_study_10/grounds.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

# Polygon
writeRaster(polygons_ras, "Data/output/Chapter6/Case_study_10/regions_grid.asc", overwrite=TRUE, gdal="DECIMAL_PRECISION=0", NAflag=-9999)

dir.create("Data/output/Chapter6/Case_study_10/output")

# Julia codes

# cd("C:\\Users\\cwbei\\Dropbox\\GitHubProjects\\Osa-Conservation-Connectivity-Project\\data\\Chapter6\\Case_study_10")
# compute("case_study_10.ini")

```


```{r, echo=F, message=F, warning=F}
# Import the original cost surface
lc <- rast("Data/output/Chapter6/Case_study_10/cost_surface_200m.tif")
aoi <- ext(lc)

# Plot the result - TIME TAKEN 6454 seconds - will try with different solver
res_cur <- rast("Data/output/Chapter6/Case_study_10/output/Case_study_10_curmap.asc")
  
end <- crop(end_all, aoi)
end_sf <- sf::st_as_sf(end)

#plot(res_cur)

res_clamp <- clamp(res_cur, upper=12)

# Convert the landcover map
# mask outlow voltage areas
res_clamp <- app(res_clamp, fun=function(x){ x[x < 0.01] <- NA; return(x)} )

# I am going to have to downscale this too
res_clamp_low <- aggregate(res_clamp, fact=2, fun="mean", na.rm=T)


#plot(res_clamp)
# Add in protect areas (cropped to the area)
tmp <- crop(start_all, aoi)
start_sf <- sf::st_as_sf(tmp)

#plot(res_clamp)
final_cur    <- as(res_clamp,"Raster")
final_cur_low    <- as(res_clamp_low,"Raster")

cost    <- as(lc,"Raster")

boundary<- st_as_sfc(st_bbox(final_cur))

pal <- colorNumeric(c("#0C2C84", "#41B6C4", "#FFFFCC"), c(values(final_cur),12.1),
  na.color = "transparent")


leaflet() %>%
  addProviderTiles("Esri.WorldImagery", group="Satellite") %>%
  addProviderTiles("OpenStreetMap", group="OSM") %>%
  addRasterImage(cost, group="Conductance") %>%
  addRasterImage(final_cur_low, group="Current", colors = pal) %>%
  addPolygons(data=end_sf, fillColor = "red",stroke = F, label=end_sf$NAME, fillOpacity = 1, group="Target area") %>% 
  addLegend(pal = pal, values = values(final_cur),
    title = "Current flow") %>%
    addPolygons(data=boundary, fillColor = "yellow",stroke = TRUE, color="red",fillOpacity = 0) %>%
   addLayersControl(
    baseGroups = c("Satellite","OSM"),
    overlayGroups = c("Conductance","Current", "Target area"))%>% 
    hideGroup("Conductance") 


# Maps for sharing
writeRaster(final_cur, "Data/output/Chapter6/Climate_hub_current_and_resistance/HondurasVolcan_current.tif", overwrite=TRUE, NAflag=-9999)
writeRaster(lc, "Data/output/Chapter6/Climate_hub_current_and_resistance/HondurasVolcan_conductance.tif", overwrite=TRUE, NAflag=-9999)



```


```{r, eval=F, echo=F}

# ## High elevation connectivity
# 
# **TO BE DONE**
# 
# 
# Here I give each fragment a unique code and use the connectivity between the
# 
# 
# In particular, the all-to-one mode can be an alternative to the pairwise mode when
# the goal is to produce a cumulative map of important connectivity areas among
# multiple source/target patches.
# 
# 
# NOT WORKING




# Here i just need the protected area fragments
# Create folders to store the results
dir.create("Data/output/Chapter6/High_connectivity_low_res")
path_text <- "Data/output/Chapter6/High_connectivity_low_res"

# read them back in with the raster package
end_polygons <- raster("Data/output/Chapter6/Coastal_pa_all_low_res/end_poly_stars.tif")
grounds_polygons <- raster("Data/output/Chapter6/Coastal_pa_all_low_res/grounds_stars.tif")
cellmap <- raster("Data/output/Chapter5/Low_res_resistance_surface.tif")


# Match resolutions with the cost surface (cellmap)
plot(cellmap)
plot(grounds_polygons, add=T, col="blue")
plot(end_polygons, add=T, col="red")

end_final       <- extend(end_polygons, cellmap)
grounds_final   <- extend(grounds_polygons, cellmap)


extent(cellmap)
extent(grounds_final)
extent(end_final)

# Give each source an ID
grounds_final[grounds_final>0] <- 1
length(grounds_final[grounds_final>0])
length(unique(end_final[end_final>0]))

# This inst working in circuitscape
#instead make the grounds from the poly file
#new_grounds <- end_final
#new_grounds[new_grounds>0]<- 1

length(new_grounds)
writeRaster(cellmap, paste0(path_text,"/regions_grid.asc"), overwrite=T)
writeRaster(end_final, paste0(path_text,"/polygons.asc"), overwrite=T)
writeRaster(grounds_final, paste0(path_text,"/points.asc"), overwrite=T)

dir.create(paste0(path_text,"/output"))



```
